Split visor init.go (3637 lines) into 6 topic files
Pure reorganization — no code changes:
- init.go (260): module registration, orchestration
- init_dmsg.go (615): DMSG client, servers, HTTP, trackers, pty
- init_transport.go (877): transport manager, STCPR/SUDPH, TPD
- init_router.go (271): router, embedded route setup, embedded TPS
- init_apps.go (569): launcher, CLI/RPC, hypervisors, ping handlers
- init_services.go (1158): health, uptime, survey, autoconnect, UI server
